\name{LCFdata-package}
\alias{LCFdata-package}
\alias{LCFdata}
\docType{package}
\title{Data sets for package \code{LMERConvenienceFunctions}}
\description{This package contains (1) event-related brain potential data recorded from 10 participants at electrodes Fz, Cz, Pz, and Oz (0--300 ms) in the context of Antoine Tremblay's PhD thesis (Tremblay, 2009); (2) ERP amplitudes at electrode Fz restricted to the 100 to 175 millisecond time window; and (3) plotting data generated from a linear mixed-effects model from Tremblay & Newman (In Preparation).}
\details{
\tabular{ll}{
Package: \tab LCFdata\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-12-12\cr
License: \tab GPL-2\cr
}
}
\author{Antoine Tremblay, Department of Psychology, Dalhousie University

Maintainer: Antoine Tremblay, \email{trea26@gmail.com}
}

\source{Tremblay, Antoine. (2009). \emph{Processing Advantages of Lexical Bundles: Evidence from Self-paced Reading, Word and Sentence Recall, and Free Recall with Event-related Brain Potential Recordings}. Ph.D. Dissertation. University of Alberta, Edmonton, Canada. Available for download at \url{http://www.ualberta.ca/~antoinet/ThesisDraft_10_B.pdf}.}

\seealso{
\code{\link{eeg}; }
\code{\link{erpFz}; }
\code{\link{z} }
}

\keyword{ package }
